Add files via upload
I have developed a video game titled "DragonFlier" wherein players assume the role of a dragon navigating through various environments in pursuit of collecting eggs. The objective is to gather these eggs when they reach optimal ripeness, indicated by their transition from white to dark blue with a red dot. Consuming ripe eggs rewards the player with one point; however, prematurely ingesting them results in a loss of life due to their poisonous nature.

As the game progresses, there is a risk of eggs hatching into hostile creatures that pose a threat to the player. While players have the option to defend themselves by shooting fireballs, there is a cooldown period between shots, during which only ineffective bursts of smoke are emitted. Currently, there are a total of 12 eggs scattered throughout the caverns, each contributing to the player's score upon successful collection. Achieving the maximum score necessitates strategic planning and exploration to uncover all eggs in the correct sequence.

To provide players with a fresh start or to reset the game, they can simply press the "R" key to initiate a restart.